Backend/requirements.txt
flask
flask-cors
tensorflow
numpy
fpdf
Pillow
werkzeug
google-generativeai